requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.4.0
openpyxl==3.1.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.1
psycopg2-binary==2.9.9
httpx==0.27.0
